+++ CHANGELOG.md
@@ -4,14 +4,20 @@ All notable changes to this project will be documented in this file.\
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines
 and [Semantic Versioning](https://semver.org/spec/v2.0.0.html) for versioning adherance.
 
+## [💾](https://github.com/kyaulabs/grimoire/releases/tag/v0.2.0) [0.2.0](https://github.com/kyaulabs/grimoire/compare/v0.1.0...v0.2.0) - (2023-05-23)
+
 ### ♻️ Refactor
 
 - [Tests] Test naming scheme and proper format ([413045e](https://github.com/kyaulabs/grimoire/commit/413045e5689255062782618fbbb797dda49409ac))
 
 ### 📝 Documentation
-- Xml documentation added ([044c71d](https://github.com/kyaulabs/grimoire/commit/044c71dceff146e47459f41716e8f8a994161055))
+- Xml documentation added ([14811ed](https://github.com/kyaulabs/grimoire/commit/14811ed0a5c78b7601498b29cced6c11d51407ee))
 
-<!-- Auto-generated on 2023-05-22T19:59:09.527074-07:00 -->
+
+### 🚀 Features
+
+- [Editor] Project window file extensions for one column layout ([d4e3c5e](https://github.com/kyaulabs/grimoire/commit/d4e3c5e7c2c441a0aff8410a6e1134a60e8cca71))
+<!-- Auto-generated on 2023-05-22T20:22:06.704513200-07:00 -->
 
 ## [💾](https://github.com/kyaulabs/grimoire/releases/tag/v0.1.0) [0.1.0](https://github.com/kyaulabs/grimoire/commit/05b03fc749e4bc3b4f3a19974564be5e3e69a021) - (2023-05-22)
 
@@ -22,4 +28,4 @@ and [Semantic Versioning](https://semver.org/spec/v2.0.0.html) for versioning ad
 ### 🚀 Features
 
 - [Editor] Automated project setup ([05b03fc](https://github.com/kyaulabs/grimoire/commit/05b03fc749e4bc3b4f3a19974564be5e3e69a021))
-<!-- Auto-generated on 2023-05-22T19:59:09.527471400-07:00 -->
+<!-- Auto-generated on 2023-05-22T20:22:06.704930100-07:00 -->